fix: mpack writer holding invalid data ptr
The pointer obtained in Data.withUnsafeBytes is only valid within the
closure. If mpack_writer_init_data is called within the closure, the
pointer passed to it will be invalid once the closure exits, which means
the writer will be holding onto an invalid pointer. This can cause issue
later when calling mpack_writer_try_parse since it can potentially be
parsing gibberish. Instead, the Data should be converted to [UInt8]
first, then we obtain the pointer to the array which is long lived. This
hopefully shouldn't cause a copy due to copy-on-write semantic.